+++ activesupport/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Rails 4.0.0 (unreleased) ##
 
+*   Dependencies no longer trigger Kernel#autoload in remove_const [fixes #8213]. *Xavier Noria*
+
 *   Simplify mocha integration and remove monkey-patches, bumping mocha to 0.13.0. *James Mead*
 
 *   `#as_json` isolates options when encoding a hash.